refactor(js-runtime): folders architecture
