+++ packages/js-runtime/src/runtime/global/console.ts
